html/semantics/scripting-1/the-script-element/script-defer.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: HTMLScriptElement - defer</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>

var flag = 0;
var NUMBER_FIRST = 3;
var NUMBER_SECOND = 5;
var EXPECTED = 8;

</script>

<script type="text/javascript" src="defer.js" defer></script>

<script>

function beforeLoad(){
  flag += NUMBER_FIRST;
  test(function(){
    assert_equals(flag, NUMBER_FIRST, "The value of variable is incorrect.");
  }, "script_defer_no");
}

beforeLoad();

</script>
